pyproject.toml
torch = "2.2.2"
torchvision = "0.17.2"
torchaudio = "2.2.2"
numpy = "1.26.4"


[tool.poetry]
name = "parrot"
version = "0.1.0"
description = ""
authors = ["JosephNK <nkw0608@gmail.com>"]
readme = "README.md"
packages = [{include = "parrot", from = "src"}] 

[tool.poetry.dependencies]
python = "^3.11"
torch = [
    {version = "2.2.2", markers = "platform_machine == 'i386' or platform_machine == 'i686' or platform_machine == 'x86_64'"},
    {version = "^2.7.1", markers = "platform_machine != 'i386' and platform_machine != 'i686' and platform_machine != 'x86_64'"}
]
torchvision = [
    {version = "0.17.2", markers = "platform_machine == 'i386' or platform_machine == 'i686' or platform_machine == 'x86_64'"},
    {version = "^0.22.1", markers = "platform_machine != 'i386' and platform_machine != 'i686' and platform_machine != 'x86_64'"}
]
torchaudio = [
    {version = "2.2.2", markers = "platform_machine == 'i386' or platform_machine == 'i686' or platform_machine == 'x86_64'"},
    {version = "^2.7.1", markers = "platform_machine != 'i386' and platform_machine != 'i686' and platform_machine != 'x86_64'"}
]
transformers = "^4.52.4"
sentencepiece = "^0.2.0"
accelerate = "^1.7.0"
huggingface-hub = "^0.33.0"
numpy = [
    {version = "1.26.4", markers = "platform_machine == 'i386' or platform_machine == 'i686' or platform_machine == 'x86_64'"},
    {version = "^2.3.0", markers = "platform_machine != 'i386' and platform_machine != 'i686' and platform_machine != 'x86_64'"}
]
pandas = "^2.3.0"
load-dotenv = "^0.1.0"
protobuf = "^6.31.1"
fastapi = "^0.115.12"
uvicorn = "^0.34.3"
redis = "^6.2.0"
faiss-cpu = [
    {version = "1.10.0", markers = "platform_machine == 'i386' or platform_machine == 'i686' or platform_machine == 'x86_64'"},
    {version = "^1.11.0", markers = "platform_machine != 'i386' and platform_machine != 'i686' and platform_machine != 'x86_64'"}
]
sentence-transformers = "^4.1.0"

[tool.poetry.group.dev.dependencies]
poethepoet = "^0.35.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
dev = "uvicorn src.parrot.api.main:app --reload"
start = "uvicorn src.parrot.api.main:app --host 0.0.0.0 --port 8000"
prod = "uvicorn src.parrot.api.main:app --host 0.0.0.0 --port 8000 --workers 4"